# Conflicts:
#	lib/View/Screens/body_detail_screen.dart

+++ lib/Model/pain_dashboard_model.dart
@@ -2,8 +2,8 @@ class PainDashboardModel {
   int? usersPainId;
   int? bodyPartId;
   String? bodyPart;
-  int? locationX;
-  int? locationY;
+  double? locationX;
+  double? locationY;
   String? description;
   bool? current;
   int? startDate;
